lava: unify interface with lamp

+++ lib/include/lava/support/base.hpp
@@ -237,7 +237,7 @@ namespace __lava
         static st lift_f32( f32 v ) { return st::lift( v ); }
         static st lift_f64( f64 v ) { return st::lift( v ); }
 
-        static st lift_arr( void *v, size_t s ) { return st::lift( array_ref{ v, s } ); }
+        static st lift_arr( void *v, unsigned s ) { return st::lift( array_ref{ v, s } ); }
         static st lift_ptr( void *v ) { return st::lift( v ); }
 
         template< typename size >
